bound requests by number of pages instead of transactions
We need an upper-bound to prevent an endless loop. Beyond this, the max number
of pages defined has no meaning.

Ideally we would like to hand control to the user so she/he can make the
call regarding whether to continue loading more pages. There are many UI
patterns we can use to implement this feature though.